Fix RabbitMQ ocf_run with the timeout command.
W/o this fix, the timeout command is used with rabbitmqctl
wrapper, which uses a 'su' when invoked as not a rabbitmq user.
This is an issue as the 'su' changes the original
process group. And if the timeout command expired, it would kill
only processes in this original process group, leaving orphaned
commands what do not belong to whis group anymore.

The solution is to issue all timeout wrapped rabbitmqctl commands
as the rabbitmq user in the OCF script.

Closes-bug: #1446241

Change-Id: I139255237fd34b555f248cb826deb13b7e134e8d
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>
(cherry picked from commit b725527d92c31373c84cd780dcf1ad10933f4955)