Fix RabbitMQ OCF stop action when no PIDFILE exists
W/o this patch, RabbitMQ OCF action stop would fail
if there is no pid file exist. This is an issue as that
could leave the beam process running and uncontrolled
by the Pacemaker resource agent.

The solution is to invoke 'rabbitmqctl stop' and return generic
error only if this operation have failed as well or timed out.
For this case, there also should be a note provided in logs about
additional manual stop actions required.

Closes-bug: #1446526

Change-Id: I86272230bd7f0ef8412ccaeb6bbb5dcaa18387bc
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>