Enable rabbit_heartbeats for all MOS services
W/o this fix AMQP heartbeats defaults are implicit and may lead
to broken state of some Openstack services, for example like
it was described here https://bugs.launchpad.net/fuel/+bug/1289200
This is an issue as it may lead to unstable failover procedure for
open AMQP conversations.

The solution is to enable explicit values for heartbeats both server
and client sided, for example for nova-conductor and nova-compute,
neutron-server and its agents, and follow recommendations from the
scale lab testing team to set rabbit_heartbeat to 520 as this
parameter should be lower then 580, which is a default timeout of
the rabbitmq-server.

Note, that rabbit_heartbeats option is
avaiable only in MOS Juno and later it must be replaced to upstream
implementation in Kilo as well.

Closes-bug: #1447559

Change-Id: I1f477f27ba5bb11b3cf6150c51f8da4e1e7fb963
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>