Enable rabbit_heartbeats for all MOS services
W/o this fix rabbit_heartbeats use inoptimal defaults
and are enabled not for all OpenStack services. This is
an issue as it may lead to unstable failover procedure for
open AMQP conversations.

The solution is to enable heartbeats both server and client sided,
for example for nova-conductor and nova-compute, for neutron-server
and all of its agents, and follow recommendations from the scale lab
testing team to set rabbit_heartbeat to 520 as this parameter should
be lower then 580, which is a default timeout of the rabbitmq-server.

Note, that rabbit_heartbeats option is
avaiable only in MOS Juno and later it could be replaced to upstream
implementation in Kilo as well.

Closes-bug: #1447559

Change-Id: I1f477f27ba5bb11b3cf6150c51f8da4e1e7fb963
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>