Use persistent names for journal devices
On systems with multiple disk controllers drive names
will not be consistent. On reboot an OSD's journal may
have changed names which will prevent the OSD from starting.
Instead we will attempt to use the persistent by-id link
and fall back to the device name (e.g. /dev/sda) if it's
not present.

Change-Id: I18f3c1cb31dedf9e759edf0094ebce1fdb42492d
Closes-bug: #1441434