Update atop logrotate configuration
The previous logrotate configuration for atop was repeatedly rotating
already rotated files.  Additionally, it was performing an extra restart
of atop which was not necessary because that is handled via a daily
cron.

This update performs a rotation on the daily.log file which is created
by atop and was previously not rotated. When this file is rotated, we
then apply our retention policy on the atop_YYYYMMDD files to ensure we
are keeping only the configured number of files. It should be noted that
the notifempty was removed because the lastaction only fires if at least
one log will actually be rotated.

Change-Id: I76e3c12b46946ed7c48719f51e953528708c2d4b
Closes-Bug: 1452389