Update atop logrotate configuration
The previous logrotate configuration for atop was configured to
rotate the binary files and was repeatedly rotating already rotated
files.  Additionally, it was performing an extra restart of atop which
was not necessary because that is handled via a daily cron.

This update will handle the rotation of any *.log files generated by
atop as necessary and moves the binary log retention to a daily cron
file.

Change-Id: I76e3c12b46946ed7c48719f51e953528708c2d4b
Closes-Bug: 1452389