Use retries for flavor-create
W/o this fix, keystone backend for tokens may
sporadically fail to find a token on each nova CLI command.
It may even fail if there would be a new token passed
as an argument for each CLI command.

For now, the solution is cannot be done at puppet side, but
only the workaround, which is to retry nova CLI commands

Related-bug: #1449584

Change-Id: I0f4895fe01c6811b89699677a46cbcecc6f04930
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>