Connectivity tests after netconfig, ntp, dns
Adding a repository connectivity tests to be run after netconfig to ensure
that software repository access is OK before proceeding with the rest of the
deployment.  Additionally, adding in post connectivity tests after the DNS
server and NTP server to ensure that those services will be able to
reach the configured settings.

With this change, additional functions have been added to the TestCommon
class to perform DNS lookups (TestCommon::Network.resolve?) and an NTP
request (TestCommon::Network.ntp?).

Change-Id: I6b0302ce403871384d377aceb7e94b09126b885e
Closes-Bug: 1261940