Adding connectivity tests to run after netconfig
Adding a set of connectivity tests to be run after netconfig to ensure
that DNS, NTP and software repository access is OK before proceeding
with the rest of the deployment.  If any item fails to funnction
correctly, the deployment should stop with an error message indicating
which item has failed.

With this change, additional functions have been added to the TestCommon
class to perform DNS lookups (TestCommon::Network.resolve?) and an NTP
request (TestCommon::Network.ntp?).

Change-Id: I6b0302ce403871384d377aceb7e94b09126b885e