Remove base os from configure_default_route
configure_default_route relies on l23network (netconfig) task has been run.
However it's not required that the base-os role has run this, so it will
result in a failed deployment. If a plugin is using base-os then it will
need to determine if l23network/netconfig and configure_default_route
need to run.

Change-Id: Ie734c4d8c1211a380405de0d8383c5d14b5d2f38
Closes-bug: #1452925