change default value for glance/vmware_api_retry_count
There is the bug with ostf tests for glance. Test fails if it tries to
do something when auth session restarts. A part of solution is to
increase timeouts for steps of ostf tests. Second one is change
defaults for vmware_api_retry_count to increase of times glance retry
on failures.

Change-Id: Ia2d2c320fbb52ed7cc77a2396c57d42e78cd1173
Partial-Bug: #1437228