Fix incorrect shell command in ceph OSD Facter function
When DEVLINKS was on the first line of udevadm output
this function would fail to correctly find a journal
device name. This would result in an invalid ceph-deploy
command line. An empty string as the journal device name
would cause the same issue and this case has been fixed
as well.

Change-Id: I785fd463fab781ac1b769fb2eee5c6abf0ef6261
Co-Authored-By: Mykola Golub <mgolub@mirantis.com>
Closes-bug: #1453959
Related-bug: #1441434