Ensure limits are applied for rabbitmq-server
Currently limits are applied only for root user, but they are not passed
on rabbitmq user (under which processes are being run).

This commit partially backports $file_limit parameter from puppetlabs/rabbitmq
module. As Debian-based distributions do not enforce limits by using pam_limits.so
by default, we also need to configure it to do so.

Change-Id: I2dd5639ba33ca7185665348ded10ed5b05228a33
Closes-Bug: 1455755