Fix local/undef var for rabbit OCF
W/o this fix some variables are missing the local
declaration and might be reused in another places
leading to unpredictable results. For example,
there is the nodelist var should be local as it is
used both action_notify(), get_monitor().

The solution is to fix missing local declarations
for variables and remove unused ones as well.

Closes-bug: #1455761

Change-Id: Iba5e9d984083acea1392cad0abd6453f5d6fbf8b
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>