Revert "Merge "Fix get_monitor function and local vars for rabbit OCF""
This reverts commit 0562e2ae27c1d6b2e027e63c5af4e1153f44224b, reversing
changes made to 5fd379f0cd8106808ab9fc098ca1094c7d91041a.

The issue is that by design, the start action will start beam process
and stop the rabbit app, unless master promoted or slave joined cluster.
We have to revert aforementioned commit as it introduced a regression
to the action monitor, which is to report "Not running", when rabbit app
is stopped.

Related-bug: #1455761
Change-Id: I686c620092909351c910598aeef0dcc73d2da080