Increase repo url availability timeout values
In order to improve the experiance when deploying nodes on slow or
congested network links, we need to increase the values of the
timeouts for our url availability check. We are increasing them from
5 seconds for the open and read timeouts to 60 seconds. Additionally
we are increasing the overall timeout on the check to 180 seconds. We
are also increasing the task timeout from 600 seconds to 3600 seconds.

Change-Id: Ibd4e87fbf3a27d1fca1bcad39ee9f534c9a92f60
Closes-Bug: 1456805
Related-Bug: 1261940