Ensure the rabbit-fence daemon respawned by upstart
W/o this fix, if the daemon crashed, it
would remain stopped. Also, the rabbit fence
daemon will crash on start, when there is no
/var/run/rabbitmq piddir exist.

The solution is to add the respawn option to
the daemon's upstart, and ensure the piddir
created by upstart before starting the daemon.
This also requires to add the rabbitmq package
to be installed as a dependency.

Closes-bug: #1456791

Change-Id: I4e71eb9e4aa4ff3b877aa89a37d82215740aaeab
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>