Fix rabbit OCF monitor check for running and clustered
W/o this fix, there is a situation possible when the action start
joins the node to the cluster (it becomes visible in db_nodes),
but the monitor action might not see, if it is absent in
the running_db_nodes list.
This is an issue as such node will never start as a cluster
member.

The fix is to check if current node is running and clustered and
return OCF_NOT_RUNNING if it is not.

Closes-bug: #1455761

Change-Id: I8798e776f291ac6de9884c5d07dacd5a4e1d42da
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>