Do lazy start attempts of the rabbit app
W/o this patch, the rabbitmq app may remain joined cluster
but not started and keep running without being noticed by
OCF logic. But by design, the app is supposed to be stopped for
some cases and, in general, the monitor action should not
treat this state as en error.

This is a problem as we don't want the rabbit app retained
stopped because in this state the rabbit node cannot process
AMQP at all.

The solution is to make the rabbit app at least have tried
to start (a lazy start attempt) in the end of monitor
and ignore results because it is normal when the app cannot start
for some transient cluster states.

Closes-bug: #1458828

Change-Id: Iffdb53860a4b53b0f7b6b1798aeb8088e7bc542c
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>