Do lazy start attempts of the rabbit app
W/o this patch, the rabbitmq app may remain joined cluster
but not started and keep running without being noticed by
OCF logic. But by design, the app is supposed to be stopped for
some cases and monitor action should not treat this state
as en error.

This is a problem as we don't want permanent states with
the rabbit app retained stopped because in this state the
rabbit node cannot process AMQP at all. Besides that,
we also want the rabbit app to be ready ASAP, but w/o some
"lazy start" attempts, the unavailability window only increases.

The solution is to make the rabbit app at least have tried
to start (a lazy start attempt) in the end of monitor
and post-promote and ignore results because it is normal
when it cannot start for some transient cluster states.
And at the post-start, also check if the rabbit app can be
started and report the resource as not running, if it cannot.
This will ensure eventual readiness of the rabbitmq app in the
stop-start loop along with minimizing its downtime.

Closes-bug: #1455761

Change-Id: Iffdb53860a4b53b0f7b6b1798aeb8088e7bc542c
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>