Add glance::api dependancy to glance cache crons
The glance cleanup crons cannot be processed before the glance api
package gets installed because the user may not exist.  This change
adds an explicit dependency on the glance::api class for the
glance::cache::pruner and cleaner classes.  With out this, puppet
would sometimes attempt to create these crons for the glance user
which had not been created yet and this would result in an error.

Change-Id: I2eb1f28dd9db520a1a40b82aa8d87489ff366920
Closes-Bug: 1459227