Allow for erlang_cookie from hiera
Currently the erlang_cookie for rabbitmq was hard coded but there
was a hiera lookup for a erlang_cookie value. This change uses
the hiera lookup to allow for erlang_cookies to be specified via
hiera but will default to the previous hardcoded value if no value
is present in hiera.

Change-Id: Id24eb49bda02efc5e9776b9441f211d96e7325e5
Closes-Bug: 1459315