Don't run dns-client & ntp-client tasks on base-os
We should not run dns-client & ntp-client tasks for base-os role,
because they setup vip_management address as dns/ntp server while we
don't have setuped interfaces and hence that vip is unreachable.

Moreover, we should not perform any "deployment" tasks on nodes
of such type. It suppose to be "a node after provisioning" and
that's it.

In order to avoid this issue, let's specify explicitly on which
roles we want to run dns-client & ntp-client tasks.

Change-Id: Ie0375df4095de6d78ef167e9af1e18a962f48829
Closes-Bug: #1456661