Make HAProxy check of swift proxy backends via management VIP
W/o this fix, when the management interface on the controller node
running a Swift proxy is down, HAProxy would fail to update
its backend status at the storage network.

This is a problem as we want swift backends not able to connect the
swift endpoint via the management VIP to be marked down. Othewise,
responces time for any requested swift commands would be drastically
longer. Simple httpcheck option cannot resolve this because the swift
healthcheck reports OK, if conntacted via the storage network.

In order to fix this, simple healthcheck script is implemented.
This script is running as xinetid service at TCP port 49001 and
checks the Swift healthcheck report via given swift admin endpoint with
connection timeout 5 seconds and reports an HTTP 200 OK, if the result is OK.
If the result was not OK or the endpoint was not reachable,
it would report an HTTP 503 Error. Expected failover time will
be around 30 seconds.

DocImpact: Reference architecture, swift failover.

Closes-bug: #1459772

Change-Id: I55a35b45257763a20f33bd47cb5c57de53558ccf
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>