Restrict access to the galera clustercheck
W/o this fix the galera clustercheck script may
be reached via http from any internal or external
network. This is a security issue.
The solution is to allow access to the clustercheck
from only the management network, localhost and IP
240.0.0.2 which is reserved for internal haproxy usage.

Note, the management network CIDR is extracted from the
network scheme endpoint address for the br-mgmt interface.
It is represented in CIDR as <IP>/<mask> and this works
as the xinetd only_from configuration directive just fine.

Related-bug: #1460623

Change-Id: I8195b8a5b2b87e27b82e067d8dac0895cb6663e7
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>