Initial Percona cluster support added
This change adds two new custom_setup_class options to the mysql class.
With this change, we now support 'percona' and 'percona_packages' to
be passed to the mysql::server class that will cause percona to be
used.

This change adds the use_percona and use_percona_packages params to
the existing galera class. The 'use_percona' option will cause the
configuration to use the Linux distrubution classes for the Percona
installation. This only works with Ubuntu. By setting both 'use_percona'
and 'use_percona_packages' to true, the puppet configuration will
use the Percona provided packages for installation. This assumes that
the Percona packages are available to the system.

Additionally tests have been added to the galera
class to ensure that expected resources are configured for both
the regular packages and the percona packages.  By default, the
existing galera configuration remains the same and assumes the
Mirantis packages.

Partial blueprint: detach-components-from-controllers

Change-Id: I972698a8b2dae4caca6101db6d48bab405e40eae
Co-Authored-By: Alex Schultz <aschultz@mirantis.com>