Initial Percona cluster support added
This change renames the anchors used by mysql to something that can be
shared with the new percona module as a drop in replacement.
Currently, this change assumes that the packages from Percona are
available for installation. You can pass a paramter to the percona
class indicating that you would like to use the distros version of the
packages but this will remove CentOS/RedHat support as 6 does not have
percona packages distributed with it at this time.

Partial blueprint: detach-components-from-controllers

Change-Id: I972698a8b2dae4caca6101db6d48bab405e40eae
Co-Authored-By: Alex Schultz <aschultz@mirantis.com>