Restart rabbit node if cannot list its queues
W/o this fix the dead end situation is possible
when the rabbit node have no free memory resources left
and the cluster blocks all publishing, by design.
But the app thinks "let's wait for the publish block have
lifted" and cannot recover.

The workaround is to monitor results of rabbitmqctl list_queues
and restart the rabbit node, if queues cannot be listed.
This is the similar logic as we have for the cases when rabbitmqctl
list_channels hangs.

Besides that, this patch makes the monitor of CHECK_LEVEL=30 to
gather the rabbit runtime stats like consumed memory by all queues,
total VMEM available, total messages in all queues and average queue
consumer utilization. This info should help to troubleshoot failures
better.

Closes-bug: #1463433

Change-Id: I91dec2d30d77b166ff9fe88109f3acdd19ce9ff9
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>