Move interface wait time to netconfig task
For a workaround for bug 1458625, a 32 second sleep was added to the
url_available function that is run as part of the connectivity_tests
task that is run after netconfig. This wait should be part of the
netconfig task where the delay is needed when bringing up a new
bridge in ubuntu.  This will help ensure that the interfaces are ready
after the netconfig task if the url_available task is removed or
tasks are run in a different order.

Change-Id: Ie715d6bf1622d33dd10b15c5bea74e05f1bba858
Related-Bug: 1458625
Closes-Bug: 1458954