Remove raise Timeout::Error from ntp_available
THIS COMMIT IS identical to Ibf2ebb80c04e4314f01b9a45e6a2169e1cde7fbf
it was created to override -1 workflow as we need to merge stuff
for 6.1 RC3

The ntp_available function should not raise a Timeout::Error here as
it should just return false so that the next server should be checked.
This update addresses the case where a Timeout::Error is being raised
rather than having all servers checked before erroring with a
Puppet::Error if no servers answer.

Change-Id: Ide224a0e5dffca6b3dcb8c5e12b732870d07b988
Closes-Bug: 1465598