Ensure radosgw service enabled on boot
Currently, we are not ensuring that the radowgw service is enabled to
start on boot. This change removes the restart_rados.sh and adds
enable_rados.pp to ensure the service is running and will be started
on boot.

Change-Id: I58bf4a20c9d08ba3e9624fd47e3b5780011c89b7
Closes-Bug: 1466552
(cherry picked from commit 23243f3c9f865bd329ed1e65ea5198f46e46b2ac)