Fix amqp_hosts for heat
W/o this patch amqp_hosts is being passed
to heat as array with single element containing
a string with all amqp hosts listed separeted by a comma.
This is wrong as by desgin, the heat module expects
them to be listed as array elements instead. This
leads to ha_queues configured always to False
breaking Heat HA operations.

The fix is to split amql_hosts before to pass it
to the heat class

Closes-bug: #1465840

Change-Id: Ief0266d468bf127156db8c42c86b840436090ab8
Signed-off-by: Bogdan Dobrelya <bdobrelia@mirantis.com>