Add 55572 to the reserved ports list for murano
This change adds 55572 to the reserved ports list for murano. Without
this change, some deployments fail occastionally when port 55572 is in
use. Because this change is in the murano task, we are setting the ports
to the defaults and 55572 as the defaults are added the first time in the
netconfig task.

Change-Id: If62504a5b3b7110d35278fe3654cdbb35dbbe194
Closes-Bug: 1467024