Fix amqp_hosts for Heat
W/o this patch amqp_hosts is being passed
to heat as array with single element containing
a string with all amqp hosts listed separeted by a comma.
This is wrong as by desgin, the heat module expects
them to be listed as array elements instead. This
leads to ha_queues configured always to False
breaking Heat HA operations.

The fix is to split amql_hosts before to pass it
to the heat class

Change-Id: Ib55cd954f0798309af6d14022ed66c926c11fbda
Closes-bug: #1465840