Split out haproxy configuration tasks
This change allows for specific haproxy tasks to be enabled and disabled
via hiera. In addition to being able to enable/disable tasks, the
splitting out of tasks allows for haproxy configuration to be moved
off of the controller and for specific items to be targeted to
alternative hosts.

Removed redundant haproxy_nodes hash which duplicates the feature
of $servicename_{names,ipaddresses} in hiera.

Partial blueprint: detach-components-from-controllers
Co-Authored-By: Alex Schultz <aschultz@mirantis.com>
Co-Authored-By: Matthew Mosesohn <mmosesohn@mirantis.com>

Change-Id: Iaea16127ba5565a418d158ca10e17db45958504d