Make predefined_networks parameter optional
Only create predefined networks in Neutron if 'predefined_network' attribute in
clster settings is present and not empty. Otherwise, skip this part of manifest.

So, if you need to skip creation of predefined networks, delete this parameter
or leave it empty in deployment settings.

It is important to note that if user disabled creation of predefined networks,
some components that depend on them might fail to install. The list
(non-comprehensive) includes:

* Murano
* Sahara
* Mellanox plugin for Neutron

Other services might be impacted as well.

Change-Id: Iceed16072628cced6de9f8704b35f21fd4342e87
Closes-bug: 1471244
DocImpact