Define galera gcache option
The gcache size defines how many writesets the donor node can serve in
IST, which is always the preferred choice when joining a node to a
cluster. When a Galera node rejoins the cluster, it will check the local
gcache to determine whether it needs to perform IST or SST. It will calculate
the gap between group state and the local state. If all of the writesets
that it missed can be found in the donor's gcache, it will perform IST
by getting the missing writesets and catch up with the group by replaying them.
This is faster than SST, and non-blocking on the donor side.

Partly implements: blueprint mos-node-reinstallation
Change-Id: I0e1c086be8369713cda4b71b18b3b104aedab285