Initial librarian-puppet-simple configuration
This change adds support to leverage a Puppetfile located in the
deployment folder to add additional upstream modules.  A new script
called update_modules.sh is now available in the deployment directory
that will run librarian-puppet-simple to update the modules.  The noop
tests have been updated to leverage this script prior to running the
tests. Additionally the fuel-library spec has been updated to execute
the update_modules.sh script as part of the rpm build process.

It should be noted that this is going to use librarian-puppet-simple
which only supports git references and tarballs and provides no
dependancy lookup. This is a much simpler way of pulling in the modules.

Partial blueprint: fuel-puppet-librarian

Change-Id: I5e628f159d2d11121af741bcc1218f292cd2b96e