Initial librarian-puppet configuration
This change adds support to leverage a Puppetfile located in the
deployment folder to add additional upstream modules.  A new script
called update_modules.sh is now available in the deployment directory
that will run librarian to update the modules.  The noop tests have been
updated to leverage this script prior to running the tests.

Partial blueprint: fuel-puppet-librarian

Change-Id: I5e628f159d2d11121af741bcc1218f292cd2b96e