cobbler: pass the correct root FS URL to bootstrap nodes
Cobbler passes wrong kernel command line to Ubuntu based bootstrap,
that is, it instructs the node to download the root filesystem image
from localhost. In nailgun/examples/cobbler-only.pp server gets set to
localhost, but then in each distro we set server variable manually to
fuel master IP. This was required during the time when we ran a NAT
for containers and they had some internal NATed IP using the internal
IP would break tftp. It can be removed now that we're on host networking
and replaced with the fuel master IP as usual.

Note: before the commit f01d4c6f570aa61d (`cobbler: pass root FS URL
to bootstrap kernel command line') the correspoding manifest was using
the $fuel_settings global variable (so the ip address was correct).

blueprint fuel-bootstrap-on-ubuntu

Change-Id: Ie62188fd95f967a954dd60a186d6ee1ff1dc9c1f