Make DNS on the master node work even when cobbler is not running
Currently resolv.conf on the master node sets the master node itself
as the only DNS server. The requests are served by dnsmasq running in
the cobbler container. Thus name resolution does not work when cobbler
is not running (not ready yet, the container is being updated, docker
is being updated, etc). This breaks updates and other actions (such as
building bootstrap images) which need a working DNS. To solve the issue
add the upstream DNS servers to resolv.conf.

blueprint fuel-bootstrap-on-ubuntu

Change-Id: I16261d2eaf37c3f59f28a494d38ee72c6d3aa2d2