Support Ceph OSD devices with existing data set
Partition preservation feature allows to preserve data on Ceph OSD device, but
later on Puppet will run 'ceph-deploy prepare' on every Ceph device in a system.
This call destroys data set on those devices.

Partition preservation is only required if the node is being reinstalled. Pass a
hiera flag 'reinstall_node' to the Ceph module to make sure that existing
prepared Ceph OSD devices won't be formatted by Puppet.

To preserve data on Ceph OSD devices through deployment process, we need to
check if the device has Ceph data and if so, skip running 'ceph-deploy prepare'
on that device. Only prepared devices must be activated to avoid deployment
failure.

We should cover following cases with this module:

1. Install new Ceph OSD node. In this case, we don't expect any devices listed
by facter to be prepared OSDs, and run prepare/activate on every device.

In this case, 'reinstall_node' is 'false'. For every device, we run 'ceph-deploy
prepare' -> 'ceph-deploy activate' with the following conditions:

* run 'ceph-deploy activate' only if the device is 'prepared'

2. Add new OSDs on existing Ceph OSD node. In this case, we expect some devices
to be active OSDs, and all other devices found by facter must be
prepared/activated.

In this case, 'reinstall_node' is 'false'. For every device, we run 'ceph-deploy
prepare' -> 'ceph-deploy activate' with the following conditions:

* run 'ceph-deploy prepare' if the device is not 'active'
* run 'ceph-deploy activate' only if the device is 'prepared'

3. Reinstall operating system on the existing Ceph OSD node without dropping OSD
data. In this case, we expect that all devices listed in facts are prepared, and
only have to be activated. We also expect that all devices are configured for
the current cluster, thus we need to stop deployment if something goes wrong
(i.e. devices cannot be activated).

In this case, 'reinstall_node' is 'true'. For every device, we run 'ceph-deploy
activate' only if the device is 'prepared'

DocImpact
Change-Id: I667fa6aab9d6f46c73bfb8ca0e267afede6049fb
Implements: blueprint partition-preservation