Support Ceph OSD devices with existing data set
Partition preservation feature allows to preserve data on Ceph OSD device, but
later on Puppet will run 'ceph-deploy prepare' on every Ceph device in a system.
This call destroys data set on those devices.

To preserve data on Ceph OSD devices through deployment process, we need to
check if the device has Ceph data and if so, skip running 'ceph-deploy prepare'
on that device. Only prepared devices must be activated to avoid deployment
failure.

It was noted that the bug #1474510 causes ceph-osd service to start
improperly, thus preventing ceph-all init script from activating all the
existing OSD devices on boot. With this bug solved, we only need to verify
that devices are prepared or activated to skip prepare stage.

Change-Id: I667fa6aab9d6f46c73bfb8ca0e267afede6049fb
Implements: blueprint partition-preservation