Support Ceph OSD devices with existing data set
Partition preservation feature allows to preserve data on Ceph OSD device,
but later on Puppet will run 'ceph-deploy prepare' on every Ceph device in
a system. This call destroys data set on those devices.

To preserve data on Ceph OSD devices through deployment process, we need to
check if the device has Ceph data and if so, skip execution of 'ceph-deploy 
prepare' on that device. Only prepared devices must be activated to avoid
deployment failure.

Following steps are made in the strict sequence to ensure protection of 
the existing data:

* execute 'udevadm trigger' to force create links for devices in /dev/disk
* execute 'ceph-disk activate-all' to force activation of all devices
  identified as ceph osds
* in ceph::osds::osd class for every ceph osd device check if it is prepared
  or active, if not, run 'ceph-deploy prepare' on it
* for every osd device, check if it is prepared (not active) and activate it

Change-Id: I667fa6aab9d6f46c73bfb8ca0e267afede6049fb
Implements: blueprint partition-preservation